Correct size of 'int32' & 'uint32'
Details

  The size of 'int32' and 'uint32' were written incorrectly, and ended
  up being 8-bytes in size.  This change rectifies the mistake.

Testing

  source ~/lrstar/lrstar/scripts/setup \
    --build-type release \
    --bod /tmp/lrstar &&  \
  cd ${LRSTAR_DIR}

  lrstar-build -j 20 clean
  lrstar-build -j 20
  ${LRSTAR_DIR}/scripts/precommit-test

  source ~/lrstar/lrstar/scripts/setup \
    --build-type debug \
    --bod /tmp/lrstar &&  \
  cd ${LRSTAR_DIR}

  lrstar-build -j 20 clean
  lrstar-build -j 20
  ${LRSTAR_DIR}/scripts/precommit-test

   lrstar-size --compare /tmp/lrstar.unmodified/release

                                                .text         .rodata           .data            .bss
     release/examples/C/C                  7K   (-14)     23K   (-70)      2K     (0)      3K     (0)
     release/examples/Calc/Calc            8K   (-14)      5K   (-70)     944     (0)      3K     (0)
     release/examples/LRK/LRK              7K   (-18)      5K   (-70)     728     (0)      2K     (0)
     release/examples/Typedef/Typedef      7K   (-18)      4K   (-70)     576     (0)      2K     (0)
     release/grammars/ANTLR/ANTLR          6K    (-2)     15K  (-126)      1K     (0)      2K     (0)
     release/grammars/C11/C11              5K     (2)     47K  (-126)      1K     (0)      2K     (0)
     release/grammars/Java1/Java1          6K    (-2)     69K  (-126)      1K     (0)      2K     (0)
     release/grammars/Java9/Java9          7K   (-12)    132K   (-84)      4K     (0)      2K     (0)
     release/grammars/Modula2/Modula2      6K    (-2)     22K  (-126)      1K     (0)      2K     (0)
     release/grammars/PLSQL/PLSQL          6K     (0)      2M  (-80K)     35K     (0)      2K     (0)
     release/grammars/Pascal/Pascal        7K   (-14)     61K   (-70)      4K     (0)      4K     (0)
     release/grammars/verilog/verilog      6K    (-2)     84K  (-126)      3K     (0)      2K     (0)
     release/src/dfa/dfa                  72K     (0)     11K     (0)      3K     (0)     67K     (0)
     release/src/lrstar/lrstar            86K     (0)     29K   (704)      6K     (0)    106K     (0)
     release/usr/local/bin/dfa            72K     (0)     11K     (0)      3K     (0)     67K     (0)
     release/usr/local/bin/lrstar         86K     (0)     29K   (704)      6K     (0)    106K     (0)

+++ src/include/lrstar_basic_defs_gcc.h
@@ -50,8 +50,8 @@ typedef signed short int   int16;
 #endif
 
 #if __SIZEOF_INT__ == 4
-typedef unsigned long int  uint32;
-typedef signed long int    int32;
+typedef unsigned int       uint32;
+typedef signed int         int32;
 #else
 #error An integer is not 32-bits with this Gcc.
 #endif
